Refactor monkey-patching so it is active during docs build
