+++ docs/manuals/characterization/t1.rst
@@ -35,6 +35,13 @@ for qubit 0.
     install qiskit-aer qiskit-ibm-runtime``.
 
 .. jupyter-execute::
+    :hide-code:
+
+    # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+    from qiskit_experiments.test.patching import patch_sampler_test_support
+    patch_sampler_test_support()
+
+.. jupyter-execute::
 
     import numpy as np
     from qiskit.qobj.utils import MeasLevel
