+++ docs/manuals/characterization/tphi.rst
@@ -26,6 +26,13 @@ From the :math:`T_1` and :math:`T_2` estimates, we compute the results for
     install qiskit-aer qiskit-ibm-runtime``.
 
 .. jupyter-execute::
+    :hide-code:
+
+    # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+    from qiskit_experiments.test.patching import patch_sampler_test_support
+    patch_sampler_test_support()
+
+.. jupyter-execute::
 
     import numpy as np
     import qiskit
