+++ docs/manuals/measurement/restless_measurements.rst
@@ -63,6 +63,13 @@ they use always starts with the qubits in the ground state.
     backend.  You can install it with ``python -m pip install qiskit-ibm-runtime``.
 
 .. jupyter-execute::
+    :hide-code:
+
+    # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+    from qiskit_experiments.test.patching import patch_sampler_test_support
+    patch_sampler_test_support()
+
+.. jupyter-execute::
 
     from qiskit_ibm_runtime.fake_provider import FakePerth
