+++ docs/manuals/verification/quantum_volume.rst
@@ -30,6 +30,13 @@ z_value = 2), and at least 100 trials have been ran.
     install qiskit-aer qiskit-ibm-runtime``.
 
 .. jupyter-execute::
+    :hide-code:
+
+    # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+    from qiskit_experiments.test.patching import patch_sampler_test_support
+    patch_sampler_test_support()
+
+.. jupyter-execute::
 
     from qiskit_experiments.framework import BatchExperiment
     from qiskit_experiments.library import QuantumVolume
