+++ docs/tutorials/custom_experiment.rst
@@ -563,6 +563,13 @@ To test our code, we first simulate a noisy backend with asymmetric readout erro
 
 
 .. jupyter-execute::
+    :hide-code:
+
+    # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+    from qiskit_experiments.test.patching import patch_sampler_test_support
+    patch_sampler_test_support()
+
+.. jupyter-execute::
 
   from qiskit_aer import AerSimulator, noise
