+++ qiskit_experiments/library/characterization/correlated_readout_error.py
@@ -77,6 +77,10 @@ class CorrelatedReadoutError(BaseExperiment):
         .. jupyter-execute::
             :hide-code:
 
+            # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+            from qiskit_experiments.test.patching import patch_sampler_test_support
+            patch_sampler_test_support()
+
             from qiskit.providers.fake_provider import GenericBackendV2
             from qiskit_aer import AerSimulator
