+++ qiskit_experiments/library/characterization/fine_frequency.py
@@ -53,6 +53,10 @@ class FineFrequency(BaseExperiment):
         .. jupyter-execute::
             :hide-code:
 
+            # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+            from qiskit_experiments.test.patching import patch_sampler_test_support
+            patch_sampler_test_support()
+
             # backend
             from qiskit_ibm_runtime.fake_provider import FakePerth
             from qiskit_aer import AerSimulator
