+++ qiskit_experiments/library/characterization/ramsey_xy.py
@@ -85,6 +85,10 @@ class RamseyXY(BaseExperiment, RestlessMixin):
         .. jupyter-execute::
             :hide-code:
 
+            # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+            from qiskit_experiments.test.patching import patch_sampler_test_support
+            patch_sampler_test_support()
+
             # backend
             from qiskit_aer import AerSimulator
             from qiskit_ibm_runtime.fake_provider import FakePerth
