+++ qiskit_experiments/library/characterization/zz_ramsey.py
@@ -129,6 +129,10 @@ class ZZRamsey(BaseExperiment):
         .. jupyter-execute::
             :hide-code:
 
+            # Temporary workaround for missing support in Qiskit and qiskit-ibm-runtime
+            from qiskit_experiments.test.patching import patch_sampler_test_support
+            patch_sampler_test_support()
+
             # backend
             from qiskit_ibm_runtime.fake_provider import FakePerth
             from qiskit_aer import AerSimulator
